refactor: make handler package
